Switched the library to use libgpiod instead of sysfs
Description: Switched button polling from sysfs to libgpiod implementation.
The previous implementation was using EPOLL to track
changes in the gpio values, but sysfs gpio values don't support
epoll, so the buttons wouldn't work this way. Using libgpiod,
the changes in gpio line values can be detected and attachted
to an interrupt. Much of the existing code had to be changed to support
this. The "direction" JSON values are now "rising" and "falling" instead of
"in" and "out" for clarity. Also, all changes in gpio values result in
an interrupt now. This way, the handler can decide what to do with that information.
This means "direction" now specifies whether a rising or falling
edge causes "asserted" to evaluate to true. "Asserted" is a boolean given
to the eventHandler.

Tests: Tested on qemu to check gpio status and configuration.
        Also used skinnyjoe hardware to ensure button presses
        were detected and that phosphor-state-manager recieved
        dbus signals. In testing, the target responded appropriately.

Signed-off-by: Brooks Keough <bkeough@nvidia.com>

Fixes JIRA: https://jirasw.nvidia.com/browse/DGXOPENBMC-4747
Fixes JIRA: https://jirasw.nvidia.com/browse/DGXOPENBMC-4322